# Tests/Financial.Api.Tests/Contract/openapi-v1.snapshot.json
{
  "components": {
    "schemas": {
...
      "InvestmentAccountCreateDTO": {
        "properties": {
          "aliases": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "isActive": {
            "type": "boolean"
          },
          "isLiability": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "isActive",
          "isLiability",
          "aliases"
        ],
        "type": "object"
      },
      "InvestmentAccountDTO": {
        "properties": {
          "aliases": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "isLiability": {
            "type": "boolean"
          },
          "latestBalance": {
            "format": "double",
            "type": "number"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "isActive",
          "isLiability",
          "aliases",
          "latestBalance"
        ],
        "type": "object"
      },
      "InvestmentAccountUpdateDTO": {
        "properties": {
          "aliases": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "isActive": {
            "type": "boolean"
          },
          "isLiability": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "isActive",
          "isLiability",
          "aliases"
        ],
        "type": "object"
      },
...
      "ProblemDetails": {
        "properties": {
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "format": "int32",
            "type": [
              "null",
              "integer"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "type": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "",
    "title": "testhost | v1",
    "version": "1.0"
  },
  "openapi": "3.1.1",
  "paths": {
...
    "/investment-accounts": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/InvestmentAccountDTO"
                  },
                  "type": "array"
                }
              }
            },
            "description": "200 OK with the full, unfiltered list of investment accounts."
          }
        },
        "summary": "Lists all investment accounts.",
        "tags": [
          "InvestmentAccounts"
        ]
      },
      "post": {
        "requestBody": {
          "content": {
            "application/*\u002Bjson": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/InvestmentAccountCreateDTO"
                  }
                ]
              }
            },
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/InvestmentAccountCreateDTO"
                  }
                ]
              }
            },
            "text/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/InvestmentAccountCreateDTO"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestmentAccountDTO"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Bad Request"
          }
        },
        "summary": "Creates a new investment account.",
        "tags": [
          "InvestmentAccounts"
        ]
      }
    },
    "/investment-accounts/{id}": {
      "delete": {
        "parameters": [
          {
            "description": "The account\u0027s identifier.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Conflict"
          }
        },
        "summary": "Deletes an investment account, when its most recent recorded balance is zero.",
        "tags": [
          "InvestmentAccounts"
        ]
      },
      "put": {
        "parameters": [
          {
            "description": "The account\u0027s identifier.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/*\u002Bjson": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/InvestmentAccountUpdateDTO"
                  }
                ]
              }
            },
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/InvestmentAccountUpdateDTO"
                  }
                ]
              }
            },
            "text/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/InvestmentAccountUpdateDTO"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestmentAccountDTO"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "summary": "Updates an investment account\u0027s name, active flag, liability flag, and aliases.",
        "tags": [
          "InvestmentAccounts"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/api/v1/financial"
    }
  ],
...
}